% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visulization.R
\name{bubble_plot}
\alias{bubble_plot}
\title{Plot bubble plot for enrichment analysis results}
\usage{
bubble_plot(data, title = NULL)
}
\arguments{
\item{data}{result}
}
\description{
Plot bubble plot for enrichment analysis results
}
