% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/func.R
\name{nonpar_test}
\alias{nonpar_test}
\title{Create table of differentially expressed tags using nonparametric method}
\usage{
nonpar_test(x, case, method = c("fisher", "Wilcoxon"))
}
\arguments{
\item{x}{a object of SpC_list class created by SpC_list function}

\item{case}{String, the case condition}

\item{method}{the method used for statistical testing}
}
\value{
a data frame with row name the same as the data_matrix, and p-value,
adjusted p-value using BH method
\itemize{
\item logFC - log2 fold change
\item PValue - p-value calculated based on selected statistical method
\item PAdjusted - adjusted p-value using BH method
}
}
\description{
Perform non-parametric test for sequencing data
}
\details{
if method='fisher' is used.
}
\section{collapse the matrix into a 2 by 2 table with}{
\tabular{lcr}{
    \tab case \tab other \cr
   gene X \tab a \tab b \cr
   Other genes \tab c \tab c \cr
}


The Fisher's exact test is applied to each row in the data_matrix

if method='Wilcoxon' is used, a two-sided Wilxoxon-sum-rank test will be used,
the p-value is calculated using approximation because it is very likely there are
ties in the hiigh-throughput data especially there are a lot of 0's
}

\references{
Ming Li (2010), Comparative Shotgun Proteomics Using Spectral Count Data and Quasi-Likelihood Modeling
}
\author{
Chunhui Gu
}
