% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statistical_testing.R
\name{one_vs_all_test}
\alias{one_vs_all_test}
\title{One-versus-all test}
\usage{
one_vs_all_test(
  x,
  method = c("fisher", "Wilcoxon"),
  logFC = 1,
  PAjusted = 0.05,
  ...
)
}
\arguments{
\item{logFC}{the log-fold-change that the returned proteins must be large than for each of the one-versus-all tests}

\item{PAjusted}{the adjusted p-value that the returned proteins must be large than for each of the one-versus-all tests}

\item{...}{additional parameter passed into nonpar_test function}

\item{y}{a object of SpC_list class which contains a data matrix and annotation table for replicates. It is created by SpC_list function}
}
\value{
a list, each element in the list is a data.frame from a one-versus-all test result. The name shows what is the case used in the test.
For example, list[\link{'diseaseX'}] is the test result of diseaseX-versus-the-other-diseases
}
\description{
this function perform a series of one-versus-all tests with respect to every category in
the y$annotation$disease
}
\author{
Chunhui Gu
}
