% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visulization.R
\name{plot_heat_map}
\alias{plot_heat_map}
\title{Plot heatmap for SpC_list object with customized setting}
\usage{
plot_heat_map(
  x,
  upper_limit = NULL,
  log_scale = TRUE,
  row_keep = NULL,
  show_rownames = F,
  show_colnames = F,
  cluster_cols = F,
  cluster_rows = T,
  ...
)
}
\arguments{
\item{x}{x}

\item{upper_limit}{x}

\item{log_scale}{x}

\item{row_keep}{x}

\item{show_rownames}{x}

\item{show_colnames}{show_colnames, cluster_cols, cluster_rows, ... pheatmap
parameters that will be used when calling pheatmap function}
}
\description{
This is an wrapper function that use pheatmap to plot heatmap
for the SpC_list object and provide additional data processing and plotting setting that
is useful for proteomics spectral count data
}
