% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visulization.R
\name{plot_venn_diagram_top_Go}
\alias{plot_venn_diagram_top_Go}
\title{Custumorized Venn diagram plot}
\usage{
plot_venn_diagram_top_Go(top_GOs, main = "Title", ...)
}
\arguments{
\item{main}{the tilte of the plot}

\item{x}{a list of character vectors whose overlappings you want to plot}
}
\description{
This is wrapper function for VennDiagram::venn.diagram function
with a preset of parameters. It will output to R viewer instead of as a file
}
