% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/project-package.R
\docType{package}
\name{protools-package}
\alias{protools}
\alias{protools-package}
\title{protools: protools: proteomics data pre-processing and analysis}
\description{
a useful package for automating proteomics workflow for Dr. Samir M. Hanash's lab
}
\author{
\strong{Maintainer}: Chunhui Gu \email{chunhui.gu@uth.tmc.edu}

}
\keyword{internal}
