% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mapping.R
\name{remove_isoform_with_suffix}
\alias{remove_isoform_with_suffix}
\title{Remove protein isoforms that have a name suffix.}
\usage{
remove_isoform_with_suffix(x, method = c("remove", "combine"))
}
\arguments{
\item{x}{a numeric matrix, with each row represents a protein and each column represents a sample}

\item{method}{how data is processed in the processing of removing isoforms.
Use \code{remove} if you want just discard isoform data. Use \code{combine} to sum data of isoform into
non-isoform entries.}
}
\description{
example, protein \code{P08181-1} is an isoform of protein \code{P08181}.
Be aware that not all isoforms are expressed with a suffix that consists of
hyphen and number: https://www.uniprot.org/help/canonical_and_isoforms.
}
\examples{
# when use `method='remove'`
dat <- data.frame(a=c(1, 2, 3, 4, 5), b=c(1, 2, 3, 4, 5), row.names = c('P1001', 'P1001-1', 'P1002', 'P1002-2', 'P1003'))
mat <- as.matrix(dat)
remove_isoform_with_suffix(mat, method='remove')

# When use `method='combine'`
remove_isoform_with_suffix(mat, method='combine')
}
\seealso{
https://www.uniprot.org/help/canonical_and_isoforms
}
