% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocessing.R
\name{total_spectral_count_norm}
\alias{total_spectral_count_norm}
\title{Title Total spectral count normalization}
\usage{
total_spectral_count_norm(x)
}
\arguments{
\item{x}{a matrix of raw counts with a element ID in as rowname}
}
\value{
a matrix of total spectral count normalized version of original input
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#stable}{\figure{lifecycle-stable.svg}{options: alt='[Stable]'}}}{\strong{[Stable]}}
The total spectral count (TSC) normalization involve three steps:
\enumerate{
\item Calculate the total number of spectra in each BioSample
\item Calculate the average number of spectra across all BioSamples
\item Multiply each spectrum count in each sample by the average count over the BioSample's total spectrum count
}
}
\examples{
count <- matrix(
  c(12, 8,
    6, 3,
    4, 3),
  nrow = 3,
  byrow = T,
  dimnames = list(c('A', 'B', 'C'), c('Rep1', 'Rep2'))
)
total_spectral_count_norm(count)

}
\references{
https://support.proteomesoftware.com/hc/en-us/articles/
115002739586-Spectrum-Count-Normalization-in-Scaffold#:~:text=The\%
20normalization\%20process\%20involves\%20three,the\%20BioSample's\%20total\%20spectrum\%20count
}
