% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocessing.R
\name{TPM}
\alias{TPM}
\title{Calculate transcription per million (TPM)}
\usage{
TPM(x, gene_length, per_count = 1e+07, na_fill = NULL)
}
\arguments{
\item{x}{a matrix of raw counts with a gene id in as row name}

\item{per_count}{}

\item{na_fill}{the values used to fill if NA is encountered. Can be a scalar
for all NA or a vector which values will be sequentially filled into NA positions
each gene represented by the gene id in each row}

\item{length}{a data.frame that contains information about gene length for}
}
\value{
the TPM normalized matrix in the same shape as x
}
\description{
Calculate transcription per million (TPM) based on given information
}
\references{
https://www.youtube.com/watch?v=TTUrtCY2k-w&t=548s
https://www.biostars.org/p/335187/
}
