% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocessing.R
\name{transcript_per_million_norm}
\alias{transcript_per_million_norm}
\title{Calculate transcription per million (TPM)}
\usage{
transcript_per_million_norm(x, gene_length, per_count = 1e+07, na_fill = NULL)
}
\arguments{
\item{x}{a matrix of raw counts with a gene ID in as rowname}

\item{per_count}{}

\item{na_fill}{the values used to fill if NA is encountered. Should be a scalar
each gene represented by the gene ID in each row}

\item{length}{a data.frame that contains information about protein length
where the first column as the protein ID and the second colmun as the length}
}
\value{
the TPM normalized matrix in the same shape as x
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#stable}{\figure{lifecycle-stable.svg}{options: alt='[Stable]'}}}{\strong{[Stable]}}
Calculate transcription per million (TPM) based on given information
}
\references{
https://www.youtube.com/watch?v=TTUrtCY2k-w&t=548s
https://www.biostars.org/p/335187/
}
