% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statistical_testing.R
\name{contrast_dependent_missing_handling}
\alias{contrast_dependent_missing_handling}
\title{create data matrix based on conditions to compared and number of valid values per condition}
\usage{
contrast_dependent_missing_handling(
  X,
  col_annotation,
  condition1,
  condition2,
  min_num_valids
)
}
\description{
create data matrix based on conditions to compared and number of valid values per condition
}
