% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mapping.R
\name{mapping}
\alias{mapping}
\title{mapping current entry ID to another entry ID based on dictionary in form of \code{tibble}}
\usage{
mapping(keys, dictionary, na.rm = TRUE)
}
\arguments{
\item{keys}{a string vector contains IDs needed to be mapped from}

\item{dictionary}{a dataframe in which the rownames are the keys and the first
column contains values that keys will be mapped to in a one-to-one manner. The first column
is used as keys and second column will be used as they mapped values. The value in first
column must be unique. NA is allow as a key but can only appear once in the first column.}
}
\value{
a string vectors that containes new IDs after mapping
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#stable}{\figure{lifecycle-stable.svg}{options: alt='[Stable]'}}}{\strong{[Stable]}}
}
\examples{
dictionary <- tibble::tibble(key=c('a', 'b', 'c'), value=c('A', 'B', 'C'))
keys <- c('a', 'a', 'c', 'b')
mapping(keys, dictionary)
}
\seealso{
tibble: https://tibble.tidyverse.org/index.html
}
