% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mapping.R
\name{rownames_mapping}
\alias{rownames_mapping}
\title{convert row names between different identification commonly used in Omics data}
\usage{
rownames_mapping(x, dictionary, map_from, map_to)
}
\arguments{
\item{x}{a numeric matrix, cannot be a data.frame since duplicated row names are not allowed in data.frame}
}
\description{
use with caution for when there are multiple-to-multiple mapping exist in the dictionary
}
